Finish moving around the settings
