+++ lib/views/settings.dart
@@ -1,30 +1,17 @@
-import 'dart:io';
-
-import 'package:bot_toast/bot_toast.dart';
-import 'package:cached_network_image/cached_network_image.dart';
-import 'package:famedlysdk/famedlysdk.dart';
-import 'package:file_picker_cross/file_picker_cross.dart';
-
-import 'package:furrychat/components/settings_themes.dart';
 import 'package:furrychat/config/app_config.dart';
-import 'package:furrychat/utils/platform_infos.dart';
+import 'package:furrychat/views/settings/settings_account.dart';
+import 'package:furrychat/views/settings/settings_chat.dart';
 import 'package:furrychat/views/settings/settings_devices.dart';
-import 'package:furrychat/views/settings/settings_ignore_list.dart';
-import 'package:furrychat/components/avatar.dart';
-import 'package:flutter/foundation.dart';
+import 'package:furrychat/views/settings/settings_encryption.dart';
+import 'package:furrychat/views/settings/settings_homeserver.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/l10n.dart';
 import 'package:furrychat/views/settings/settings_themes.dart';
-import 'package:image_picker/image_picker.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import 'package:furrychat/components/adaptive_page_layout.dart';
-import 'package:furrychat/components/content_banner.dart';
-import 'package:furrychat/components/dialogs/simple_dialogs.dart';
 import 'package:furrychat/components/matrix.dart';
 import 'package:furrychat/utils/app_route.dart';
-import 'package:furrychat/views/settings/app_info.dart';
-import 'package:furrychat/views/chat_list.dart';
 import 'package:furrychat/views/settings/settings_emotes.dart';
 
 class SettingsView extends StatelessWidget {
@@ -50,181 +37,6 @@ class Settings extends StatefulWidget {
 class _SettingsState extends State<Settings> {
   Future<dynamic> profileFuture;
   dynamic profile;
-  Future<bool> crossSigningCachedFuture;
-  bool crossSigningCached;
-  Future<bool> megolmBackupCachedFuture;
-  bool megolmBackupCached;
-
-  void logoutAction(BuildContext context) async {
-    if (await SimpleDialogs(context).askConfirmation() == false) {
-      return;
-    }
-    var matrix = Matrix.of(context);
-    await SimpleDialogs(context)
-        .tryRequestWithLoadingDialog(matrix.client.logout());
-  }
-
-  void _changePasswordAccountAction(BuildContext context) async {
-    final oldPassword = await SimpleDialogs(context).enterText(
-      password: true,
-      titleText: L10n.of(context).pleaseEnterYourPassword,
-    );
-    if (oldPassword == null) return;
-    final newPassword = await SimpleDialogs(context).enterText(
-      password: true,
-      titleText: L10n.of(context).chooseAStrongPassword,
-    );
-    if (newPassword == null) return;
-    await SimpleDialogs(context).tryRequestWithLoadingDialog(
-      Matrix.of(context)
-          .client
-          .changePassword(newPassword, oldPassword: oldPassword),
-    );
-    BotToast.showText(text: L10n.of(context).passwordHasBeenChanged);
-  }
-
-  void _deleteAccountAction(BuildContext context) async {
-    if (await SimpleDialogs(context).askConfirmation(
-          titleText: L10n.of(context).warning,
-          contentText: L10n.of(context).deactivateAccountWarning,
-          dangerous: true,
-        ) ==
-        false) {
-      return;
-    }
-    if (await SimpleDialogs(context).askConfirmation(dangerous: true) ==
-        false) {
-      return;
-    }
-    final password = await SimpleDialogs(context).enterText(
-      password: true,
-      titleText: L10n.of(context).pleaseEnterYourPassword,
-    );
-    if (password == null) return;
-    await SimpleDialogs(context).tryRequestWithLoadingDialog(
-      Matrix.of(context).client.deactivateAccount(auth: {
-        'type': 'm.login.password',
-        'user': Matrix.of(context).client.userID,
-        'password': password,
-      }),
-    );
-  }
-
-  void setJitsiInstanceAction(BuildContext context) async {
-    var jitsi = await SimpleDialogs(context).enterText(
-      titleText: L10n.of(context).editJitsiInstance,
-      hintText: Matrix.of(context).jitsiInstance,
-      labelText: L10n.of(context).editJitsiInstance,
-    );
-    if (jitsi == null) return;
-    if (!jitsi.endsWith('/')) {
-      jitsi += '/';
-    }
-    final matrix = Matrix.of(context);
-    await matrix.store.setItem('chat.fluffy.jitsi_instance', jitsi);
-    matrix.jitsiInstance = jitsi;
-  }
-
-  void setDisplaynameAction(BuildContext context) async {
-    final displayname = await SimpleDialogs(context).enterText(
-      titleText: L10n.of(context).editDisplayname,
-      hintText:
-          profile?.displayname ?? Matrix.of(context).client.userID.localpart,
-      labelText: L10n.of(context).enterAUsername,
-    );
-    if (displayname == null) return;
-    final matrix = Matrix.of(context);
-    final success = await SimpleDialogs(context).tryRequestWithLoadingDialog(
-      matrix.client.setDisplayname(matrix.client.userID, displayname),
-    );
-    if (success != false) {
-      setState(() {
-        profileFuture = null;
-        profile = null;
-      });
-    }
-  }
-
-  void setAvatarAction(BuildContext context) async {
-    MatrixFile file;
-    if (PlatformInfos.isMobile) {
-      final result = await ImagePicker().getImage(
-          source: ImageSource.gallery,
-          imageQuality: 50,
-          maxWidth: 1600,
-          maxHeight: 1600);
-      if (result == null) return;
-      file = MatrixFile(
-        bytes: await result.readAsBytes(),
-        name: result.path,
-      );
-    } else {
-      final result =
-          await FilePickerCross.importFromStorage(type: FileTypeCross.image);
-      if (result == null) return;
-      file = MatrixFile(
-        bytes: result.toUint8List(),
-        name: result.fileName,
-      );
-    }
-    final matrix = Matrix.of(context);
-    final success = await SimpleDialogs(context).tryRequestWithLoadingDialog(
-      matrix.client.setAvatar(file),
-    );
-    if (success != false) {
-      setState(() {
-        profileFuture = null;
-        profile = null;
-      });
-    }
-  }
-
-  Future<void> requestSSSSCache(BuildContext context) async {
-    final handle = Matrix.of(context).client.encryption.ssss.open();
-    final str = await SimpleDialogs(context).enterText(
-      titleText: L10n.of(context).askSSSSCache,
-      hintText: L10n.of(context).passphraseOrKey,
-      password: true,
-    );
-    if (str != null) {
-      SimpleDialogs(context).showLoadingDialog(context);
-      // make sure the loading spinner shows before we test the keys
-      await Future.delayed(Duration(milliseconds: 100));
-      var valid = false;
-      try {
-        handle.unlock(recoveryKey: str);
-        valid = true;
-      } catch (e, s) {
-        debugPrint('Couldn\'t use recovery key: ' + e.toString());
-        debugPrint(s.toString());
-        try {
-          handle.unlock(passphrase: str);
-          valid = true;
-        } catch (e, s) {
-          debugPrint('Couldn\'t use recovery passphrase: ' + e.toString());
-          debugPrint(s.toString());
-          valid = false;
-        }
-      }
-      await Navigator.of(context)?.pop();
-      if (valid) {
-        await handle.maybeCacheAll();
-        await SimpleDialogs(context).inform(
-          contentText: L10n.of(context).cachedKeys,
-        );
-        setState(() {
-          crossSigningCachedFuture = null;
-          crossSigningCached = null;
-          megolmBackupCachedFuture = null;
-          megolmBackupCached = null;
-        });
-      } else {
-        await SimpleDialogs(context).inform(
-          contentText: L10n.of(context).incorrectPassphraseOrKey,
-        );
-      }
-    }
-  }
 
   @override
   Widget build(BuildContext context) {
@@ -233,16 +45,7 @@ class _SettingsState extends State<Settings> {
       if (mounted) setState(() => profile = p);
       return p;
     });
-    crossSigningCachedFuture ??=
-        client.encryption.crossSigning.isCached().then((c) {
-      if (mounted) setState(() => crossSigningCached = c);
-      return c;
-    });
-    megolmBackupCachedFuture ??=
-        client.encryption.keyManager.isCached().then((c) {
-      if (mounted) setState(() => megolmBackupCached = c);
-      return c;
-    });
+
     return Scaffold(
       body: NestedScrollView(
         headerSliverBuilder: (BuildContext context, bool innerBoxIsScrolled) =>
@@ -258,34 +61,38 @@ class _SettingsState extends State<Settings> {
           children: <Widget>[
             ListTile(
               leading: CircleAvatar(
-                backgroundColor: Theme.of(context).scaffoldBackgroundColor,
-                foregroundColor: Colors.grey,
-                child: Icon(
-                  Icons.person_outlined,
-                ),
-              ),
-              title: Text(profile?.displayname ?? 'Your account'),
+                  backgroundColor: Colors.transparent,
+                  foregroundColor: Colors.grey,
+                  child: Icon(
+                    Icons.person_outlined,
+                  )),
+              title: Text(profile?.displayname ?? L10n.of(context).account),
               subtitle: Text(client.userID),
               onTap: () => Navigator.of(context).push(
                 AppRoute.defaultRoute(
                   context,
-                  AppInfoView(),
+                  AccountSettingsView(),
                 ),
               ),
             ),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
                     Icons.dns_outlined,
                   )),
-              title: Text('Your homeserver'),
+              title: Text(L10n.of(context).homeserver),
               subtitle: Text(client.homeserver.host),
+              onTap: () async =>
+                  await Navigator.of(context).push(AppRoute.defaultRoute(
+                context,
+                HomeserverSettingsView(),
+              )),
             ),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
                     Icons.color_lens_outlined,
@@ -299,7 +106,22 @@ class _SettingsState extends State<Settings> {
             ),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
+                  foregroundColor: Colors.grey,
+                  child: Icon(
+                    Icons.chat_outlined,
+                  )),
+              title: Text(L10n.of(context).chat),
+              onTap: () async => await Navigator.of(context).push(
+                AppRoute.defaultRoute(
+                  context,
+                  ChatSettingsView(),
+                ),
+              ),
+            ),
+            ListTile(
+              leading: CircleAvatar(
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
                     Icons.insert_emoticon_outlined,
@@ -314,16 +136,22 @@ class _SettingsState extends State<Settings> {
             ),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
                     Icons.lock_outline,
                   )),
               title: Text(L10n.of(context).encryption),
+              onTap: () async => await Navigator.of(context).push(
+                AppRoute.defaultRoute(
+                  context,
+                  EncryptionSettingsView(),
+                ),
+              ),
             ),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
                     Icons.devices_other_outlined,
@@ -339,230 +167,31 @@ class _SettingsState extends State<Settings> {
             Divider(thickness: 1),
             ListTile(
               leading: CircleAvatar(
-                  backgroundColor: Theme.of(context).scaffoldBackgroundColor,
+                  backgroundColor: Colors.transparent,
                   foregroundColor: Colors.grey,
                   child: Icon(
-                    Icons.info_outline,
+                    Icons.help_outline_outlined,
                   )),
-              title: Text(L10n.of(context).about),
-            ),
-
-            /*
-            Divider(thickness: 1),
-            ListTile(
-              title: Text(
-                L10n.of(context).chat,
-                style: TextStyle(
-                  color: Theme.of(context).primaryColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-            ListTile(
-              title: Text(L10n.of(context).renderRichContent),
-              trailing: Switch(
-                value: Matrix.of(context).renderHtml,
-                activeColor: Theme.of(context).primaryColor,
-                onChanged: (bool newValue) async {
-                  Matrix.of(context).renderHtml = newValue;
-                  await Matrix.of(context)
-                      .store
-                      .setItem('chat.fluffy.renderHtml', newValue ? '1' : '0');
-                  setState(() => null);
-                },
-              ),
-            ),
-            
-            Divider(thickness: 1),
-            ListTile(
-              title: Text(
-                L10n.of(context).account,
-                style: TextStyle(
-                  color: Theme.of(context).primaryColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-            ListTile(
-              trailing: Icon(Icons.edit),
-              title: Text(L10n.of(context).editDisplayname),
-              subtitle: Text(profile?.displayname ?? client.userID.localpart),
-              onTap: () => setDisplaynameAction(context),
-            ),
-            ListTile(
-              trailing: Icon(Icons.phone),
-              title: Text(L10n.of(context).editJitsiInstance),
-              subtitle: Text(Matrix.of(context).jitsiInstance),
-              onTap: () => setJitsiInstanceAction(context),
-            ),
-
-            ListTile(
-              trailing: Icon(Icons.block),
-              title: Text(L10n.of(context).ignoredUsers),
-              onTap: () async => await Navigator.of(context).push(
-                AppRoute.defaultRoute(
-                  context,
-                  SettingsIgnoreListView(),
-                ),
-              ),
-            ),
-            ListTile(
-              trailing: Icon(Icons.account_circle),
-              title: Text(L10n.of(context).accountInformation),
-              onTap: () => Navigator.of(context).push(
-                AppRoute.defaultRoute(
-                  context,
-                  AppInfoView(),
-                ),
-              ),
-            ),
-            Divider(thickness: 1),
-            ListTile(
-              trailing: Icon(Icons.vpn_key),
-              title: Text(
-                'Change password',
-              ),
-              onTap: () => _changePasswordAccountAction(context),
-            ),
-            ListTile(
-              trailing: Icon(Icons.exit_to_app),
-              title: Text(L10n.of(context).logout),
-              onTap: () => logoutAction(context),
-            ),
-            ListTile(
-              trailing: Icon(Icons.delete_forever),
-              title: Text(
-                L10n.of(context).deleteAccount,
-                style: TextStyle(color: Colors.red),
-              ),
-              onTap: () => _deleteAccountAction(context),
-            ),
-            Divider(thickness: 1),
-            ListTile(
-              title: Text(
-                L10n.of(context).encryption,
-                style: TextStyle(
-                  color: Theme.of(context).primaryColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-            ListTile(
-              trailing: Icon(Icons.compare_arrows),
-              title: Text(client.encryption.crossSigning.enabled
-                  ? L10n.of(context).crossSigningEnabled
-                  : L10n.of(context).crossSigningDisabled),
-              subtitle: client.encryption.crossSigning.enabled
-                  ? Text(client.isUnknownSession
-                      ? L10n.of(context).unknownSessionVerify
-                      : L10n.of(context).sessionVerified +
-                          ', ' +
-                          (crossSigningCached == null
-                              ? '⌛'
-                              : (crossSigningCached
-                                  ? L10n.of(context).keysCached
-                                  : L10n.of(context).keysMissing)))
-                  : null,
-              onTap: () async {
-                if (!client.encryption.crossSigning.enabled) {
-                  await SimpleDialogs(context).inform(
-                    contentText: L10n.of(context).noCrossSignBootstrap,
-                  );
-                  return;
-                }
-                if (client.isUnknownSession) {
-                  final str = await SimpleDialogs(context).enterText(
-                    titleText: L10n.of(context).askSSSSVerify,
-                    hintText: L10n.of(context).passphraseOrKey,
-                    password: true,
-                  );
-                  if (str != null) {
-                    SimpleDialogs(context).showLoadingDialog(context);
-                    // make sure the loading spinner shows before we test the keys
-                    await Future.delayed(Duration(milliseconds: 100));
-                    var valid = false;
-                    try {
-                      await client.encryption.crossSigning
-                          .selfSign(recoveryKey: str);
-                      valid = true;
-                    } catch (_) {
-                      try {
-                        await client.encryption.crossSigning
-                            .selfSign(passphrase: str);
-                        valid = true;
-                      } catch (_) {
-                        valid = false;
-                      }
-                    }
-                    await Navigator.of(context)?.pop();
-                    if (valid) {
-                      await SimpleDialogs(context).inform(
-                        contentText: L10n.of(context).verifiedSession,
-                      );
-                      setState(() {
-                        crossSigningCachedFuture = null;
-                        crossSigningCached = null;
-                        megolmBackupCachedFuture = null;
-                        megolmBackupCached = null;
-                      });
-                    } else {
-                      await SimpleDialogs(context).inform(
-                        contentText: L10n.of(context).incorrectPassphraseOrKey,
-                      );
-                    }
-                  }
-                }
-                if (!(await client.encryption.crossSigning.isCached())) {
-                  await requestSSSSCache(context);
-                }
-              },
-            ),
-            ListTile(
-              trailing: Icon(Icons.wb_cloudy),
-              title: Text(client.encryption.keyManager.enabled
-                  ? L10n.of(context).onlineKeyBackupEnabled
-                  : L10n.of(context).onlineKeyBackupDisabled),
-              subtitle: client.encryption.keyManager.enabled
-                  ? Text(megolmBackupCached == null
-                      ? '⌛'
-                      : (megolmBackupCached
-                          ? L10n.of(context).keysCached
-                          : L10n.of(context).keysMissing))
-                  : null,
-              onTap: () async {
-                if (!client.encryption.keyManager.enabled) {
-                  await SimpleDialogs(context).inform(
-                    contentText: L10n.of(context).noMegolmBootstrap,
-                  );
-                  return;
-                }
-                if (!(await client.encryption.keyManager.isCached())) {
-                  await requestSSSSCache(context);
-                }
-              },
-            ),
-            Divider(thickness: 1),
-            ListTile(
-              title: Text(
-                L10n.of(context).about,
-                style: TextStyle(
-                  color: Theme.of(context).primaryColor,
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-            ),
-            ListTile(
-              trailing: Icon(Icons.help),
               title: Text(L10n.of(context).help),
               onTap: () => launch(AppConfig.supportUrl),
             ),
             ListTile(
-              trailing: Icon(Icons.privacy_tip_rounded),
+              leading: CircleAvatar(
+                  backgroundColor: Colors.transparent,
+                  foregroundColor: Colors.grey,
+                  child: Icon(
+                    Icons.privacy_tip_outlined,
+                  )),
               title: Text(L10n.of(context).privacy),
               onTap: () => launch(AppConfig.privacyUrl),
             ),
             ListTile(
-              trailing: Icon(Icons.link),
+              leading: CircleAvatar(
+                  backgroundColor: Colors.transparent,
+                  foregroundColor: Colors.grey,
+                  child: Icon(
+                    Icons.link_outlined,
+                  )),
               title: Text(L10n.of(context).license),
               onTap: () => showLicensePage(
                 context: context,
@@ -572,10 +201,15 @@ class _SettingsState extends State<Settings> {
               ),
             ),
             ListTile(
-              trailing: Icon(Icons.code),
+              leading: CircleAvatar(
+                  backgroundColor: Colors.transparent,
+                  foregroundColor: Colors.grey,
+                  child: Icon(
+                    Icons.code_outlined,
+                  )),
               title: Text(L10n.of(context).sourceCode),
               onTap: () => launch(AppConfig.sourceCodeUrl),
-            ),*/
+            ),
           ],
         ),
       ),
